Removed findNumericNodeHandleStack and renamed findNumericNodeHandleFiber => findNumericNodeHandle
