+++ src/renderers/native/findNumericNodeHandle.js
@@ -6,7 +6,7 @@
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  *
- * @providesModule findNumericNodeHandleStack
+ * @providesModule findNumericNodeHandle
  * @flow
  */
 'use strict';
@@ -18,12 +18,12 @@ var findNodeHandle = require('findNodeHandle');
  * The injected findNodeHandle() strategy returns the instance wrapper though.
  * See NativeMethodsMixin#setNativeProps for more info on why this is done.
  */
-module.exports = function findNumericNodeHandleStack(
+module.exports = function findNumericNodeHandleFiber(
   componentOrHandle: any,
 ): ?number {
-  const nodeHandle = findNodeHandle(componentOrHandle);
-  if (nodeHandle == null || typeof nodeHandle === 'number') {
-    return nodeHandle;
+  const instance: any = findNodeHandle(componentOrHandle);
+  if (instance == null || typeof instance === 'number') {
+    return instance;
   }
-  return nodeHandle.getHostNode();
+  return instance._nativeTag;
 };
